Move src/components/collectionManagement to src/apps/mydb/collections/
