+++ app/packs/src/apps/mydb/collections/CollectionTree.js
@@ -3,7 +3,7 @@ import { Button, OverlayTrigger, Badge, Glyphicon, Tooltip } from 'react-bootstr
 import update from 'immutability-helper';
 import CollectionStore from 'src/stores/alt/stores/CollectionStore';
 import CollectionActions from 'src/stores/alt/actions/CollectionActions';
-import CollectionSubtree from 'src/components/collectionManagement/CollectionSubtree';
+import CollectionSubtree from 'src/apps/mydb/collections/CollectionSubtree';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import InboxActions from 'src/stores/alt/actions/InboxActions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
@@ -11,7 +11,7 @@ import UIStore from 'src/stores/alt/stores/UIStore';
 import ElementStore from 'src/stores/alt/stores/ElementStore';
 import InboxStore from 'src/stores/alt/stores/InboxStore';
 import Xdiv from 'src/components/extra/CollectionTreeXdiv';
-import UserInfos from 'src/components/collectionManagement/UserInfos';
+import UserInfos from 'src/apps/mydb/collections/UserInfos';
 
 import DeviceBox from 'src/components/inbox/DeviceBox';
 import UnsortedBox from 'src/components/inbox/UnsortedBox';
