+++ app/packs/src/apps/mydb/collections/MySharedCollections.js
