+++ app/packs/src/apps/mydb/collections/SharedWithMeCollections.js
