+++ app/packs/src/apps/mydb/collections/UserInfoIcon.js
